% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmat.R
\name{enrichmat}
\alias{enrichmat}
\title{Enrichment Heatmap}
\usage{
enrichmat(
  x,
  n_top = 15,
  top_pathways = NULL,
  rownames_column = "pathway",
  NES_column = "NES",
  padj_column = "padj",
  padj_legend_title = padj_column,
  padj_cutoff = 0.05,
  plot_sig_only = TRUE,
  padj_fill = "grey",
  colors = c("#3366ff", "darkred"),
  scale_by = c("row", "column", "none"),
  cell_size = unit(14, "points"),
  filename = character(0),
  height = 5,
  width = 5,
  units = "in",
  heatmap_args = list(),
  padj_args = list(),
  save_args = list(),
  draw_args = list()
)
}
\arguments{
\item{x}{object of class \code{data.frame}; output of
\code{\link[fgsea]{fgsea}} or \code{\link{fgsea2}}. Must contain a
"contrast" column.}

\item{n_top}{integer; number of pathways to display. Defaults to the top 15
pathways that are most significantly enriched across all contrasts.}

\item{top_pathways}{character; vector of specific pathways to display. If
\code{NULL} (default), the \code{n_top} pathways will be displayed instead.}

\item{rownames_column}{character; the name of a column in \code{x} containing
unique identifiers that will be used as the rownames in the heatmap.
Default is "pathway".}

\item{NES_column}{similar to \code{rownames_column}. The name of a column
containing the normalized enrichment scores (NES) that determines the
heatmap body colors. Values between -1 and +1 (noise) will appear white.}

\item{padj_column}{similar to \code{rownames_column}. The name of a column
containing the adjusted p-values that determine the area of each circle in
the heatmap.}

\item{padj_legend_title}{character; title of the background fill legend.
Defaults to \code{padj_column}.}

\item{padj_cutoff}{numeric; cutoff for terms to be statistically significant.
If \code{plot_sig_only=TRUE}, only those pathways with at least one
\code{padj_column} value less than this threshold may appear in the
heatmap. Default is 0.05.}

\item{plot_sig_only}{logical; whether to plot only those \code{n_top} terms
that have at least one \code{padj_column} value less than
\code{padj_cutoff}.}

\item{padj_fill}{character; the background color used for values in
\code{padj_column} that are less than \code{padj_cutoff}. Default is
"lightgrey".}

\item{colors}{character; vector of length 2 specifying the colors for the
largest negative and largest positive values, respectively. Defaults to
blue (#3366ff) and red (specifically, darkred).}

\item{scale_by}{character; whether to scale the circles such that the
most-significant term in each row (\code{scale_by="row"}), column
(\code{scale_by="column"}), or overall (\code{scale_by="none"}) is of
maximum area. Default is "row" to better visualize patterns across
contrasts.}

\item{cell_size}{\code{unit} object; the size of each heatmap cell. Default
is \code{unit(14, "points")}.}

\item{filename}{character; the filename used to save the heatmap. If
\code{character(0)} (default), the heatmap will be displayed instead.}

\item{height}{numeric; height of the file in \code{units}.}

\item{width}{same as \code{height}.}

\item{units}{character; units that define \code{height} and \code{width}.
Defaults to "in" (inches).}

\item{heatmap_args}{list; additional arguments passed to
\code{\link[ComplexHeatmap]{Heatmap}}.}

\item{padj_args}{list; additional arguments passed to
\code{\link[ComplexHeatmap]{Legend}}. Modifies the adjusted p-value legend.}

\item{save_args}{list; additional arguments passed to the graphics device.
See \code{\link[grDevices]{png}} for options.}

\item{draw_args}{list; additional arguments passed to
\code{\link[ComplexHeatmap]{draw-HeatmapList-method}}.}
}
\description{
Creates a heatmap summarizing FGSEA results.
}
\seealso{
\code{\link[ComplexHeatmap]{ComplexHeatmap-package}}
}
\author{
Tyler Sagendorf
}
