% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea2.R
\name{fgsea2}
\alias{fgsea2}
\title{Fast Gene Set Enrichment Analysis}
\usage{
fgsea2(
  pathways,
  stats,
  gene_column = "entrez_gene",
  adjust.method = "BH",
  adjust.globally = TRUE,
  seed = NULL,
  ...
)
}
\arguments{
\item{pathways}{output of \code{\link{msigdbr2}}.}

\item{stats}{output of \code{\link{get_ranking}}.}

\item{gene_column}{character string; the name of a column in \code{pathways}
containing the genes in each pathway.}

\item{adjust.method}{character string; the p-value correction method. Can be
abbreviated. See \code{\link[stats]{p.adjust.methods}}.}

\item{adjust.globally}{logical; should p-values from all contrasts be
adjusted together using \code{adjust.method}? Set to \code{FALSE} if the contrasts
being tested are not closely related.}

\item{seed}{numeric or \code{NULL}; passed to \code{set.seed}.}

\item{...}{additional arguments passed to
\code{\link[fgsea]{fgseaMultilevel}}.}
}
\value{
A \code{\link[data.table]{data.table}} with results for each
  contrast. See \code{\link[fgsea]{fgsea}} for more details.
}
\description{
Custom wrapper for \code{\link[fgsea]{fgsea}} for use with
  output of \code{\link{msigdbr2}}.
}
\author{
Tyler Sagendorf
}
