% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fora2.R
\name{fora2}
\alias{fora2}
\title{Over-representation analysis}
\usage{
fora2(
  pathways,
  genes,
  universe,
  gene_column = "entrez_gene",
  minSize = 1,
  maxSize = Inf,
  adjust.method = "scale",
  adjust.globally = TRUE
)
}
\arguments{
\item{pathways}{output of \code{link{msigdbr2}}.}

\item{genes}{named list of genes.}

\item{universe}{character vector of background genes.}

\item{gene_column}{character; the name of the column in pathways containing
the elements of each gene set.}

\item{minSize}{integer; minimum size of gene sets allowed for testing.}

\item{maxSize}{integer; maximum size of gene sets allowed for testing.}

\item{adjust.method}{character string; p-value correction method. Either
"scale" (default), which will compute scaled p-values or one of
\code{\link[stats]{p.adjust.methods}}.}

\item{adjust.globally}{logical; should p-values from all contrasts be
adjusted together using \code{adjust.method}? Set to \code{FALSE} if the
contrasts being tested are not closely related.}
}
\description{
Custom wrapper for \code{\link[fgsea]{fora}}.
}
\author{
Tyler Sagendorf
}
