% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_signif.R
\name{label_signif}
\alias{label_signif}
\title{Label p-values according to significance}
\usage{
label_signif(pvals)
}
\arguments{
\item{pvals}{numeric; vector of p-values.}
}
\value{
character vector of labels: p > 0.05 = "", 0.01 <= p < 0.05 = "\\emph{",
0.001 <= p < 0.01 = "\}\\emph{", p < 0.001 = "\}\\emph{\}"
}
\description{
Use asterisks to indicate level of statistical significance.
}
