% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motrpacWAT-package.R
\docType{package}
\name{motrpacWAT-package}
\alias{motrpacWAT}
\alias{motrpacWAT-package}
\title{motrpacWAT: Helper Functions for the motrpacWATData Package}
\description{
A collection of functions for the analysis and visualization of the MoTrPAC PASS1B training study rat white adipose tissue (WAT) data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pnnl-comp-mass-spec.github.io/motrpacWAT/}
  \item \url{https://github.com/PNNL-Comp-Mass-Spec/motrpacWAT/}
}

}
\author{
\strong{Maintainer}: Tyler Sagendorf \email{tyler.sagendorf@pnnl.gov} (\href{https://orcid.org/0000-0003-1552-4870}{ORCID})

Other contributors:
\itemize{
  \item Zhenxin Hou (\href{https://orcid.org/0000-0002-1301-0799}{ORCID}) (Provided WGCNA-related code.) [contributor]
  \item Vladislav Petyuk (\href{https://orcid.org/0000-0003-4076-151X}{ORCID}) (Provided foundation for limma_full function.) [contributor]
}

}
\keyword{internal}
