% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdbr2.R
\name{msigdbr2}
\alias{msigdbr2}
\title{Get gene sets / pathways from MSigDB}
\usage{
msigdbr2(
  species = "Homo sapiens",
  genes = c("gene_symbol", "entrez_gene", "ensembl_gene"),
  gs_subcat,
  capitalize = FALSE
)
}
\arguments{
\item{species}{character; scientific or common name of species. Default is
"Homo sapiens". \code{\link[msigdbr]{msigdbr_species}} displays all
possible choices.}

\item{genes}{character; type of gene ID to fetch. Options are "gene_symbol",
"entrez_gene", and "ensembl_gene".}

\item{gs_subcat}{character vector of one or more MSigDB subcategories. See
\code{\link[msigdbr]{msigdbr_collections}} for details.}

\item{capitalize}{logical; whether to capitalize the first letter of each
description if the first word does not contain a mix of capital and
lowercase letters. Improves appearance of plots, such as those produced by
\code{\link{enrichmat}}.}
}
\value{
A \code{\link[data.table]{data.table}} with columns
"gs_exact_source" (term ID), "gs_subcat" (subcategory), "gs_description"
(term description), and a list column of gene IDs specified by \code{genes}.
}
\description{
Wrapper around \code{\link[msigdbr]{msigdbr}}. Reformats pathway
data for use with \code{\link{fgsea2}}.
}
\examples{
x <- msigdbr2(species = "rat",
              genes = "gene_symbol",
              gs_subcat = c("CP:REACTOME", "GO:MF"))
}
\references{
Liberzon, A., Birger, C., Thorvaldsdóttir, H., Ghandi, M.,
Mesirov, J. P., & Tamayo, P. (2015). The Molecular Signatures Database
(MSigDB) hallmark gene set collection. \emph{Cell systems, 1}(6), 417–425.
\url{https://doi.org/10.1016/j.cels.2015.12.004}

Dolgalev, I. (2022). msigdbr: MSigDB Gene Sets for Multiple Organisms in a
Tidy Data Format. R package version 7.5.1,
\url{https:://igordot.github.io/msigdbr}
}
