% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_analyte.R
\name{plot_analyte}
\alias{plot_analyte}
\title{Visualize clinical analyte data}
\usage{
plot_analyte(x, analyte = "")
}
\arguments{
\item{x}{a `data.frame` with required columns value, sex, and timepoint.}

\item{analyte}{character; the name of a column of `x`.}
}
\description{
Generate a scatterplot of clinical analyte values.
}
