% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_baseline.R
\name{plot_baseline}
\alias{plot_baseline}
\title{Plot baseline (pre-training) measures}
\usage{
plot_baseline(x, response, conf, stats, bracket.nudge.y = 0)
}
\arguments{
\item{x}{a \code{data.frame} with columns \code{"sex"}, \code{"age"},
\code{"group"}, and a column specified by \code{response}.}

\item{response}{character; the name of a column in \code{x} to plot.}

\item{conf}{a \code{data.frame} with confidence interval data.}

\item{stats}{optional \code{data.frame} with results of statistical analyses
filtered to the response of interest.}

\item{bracket.nudge.y}{numeric;}
}
\description{
Plot baseline (pre-training) measures.
}
