% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eigenfeature.R
\name{plot_eigenfeature}
\alias{plot_eigenfeature}
\title{Visualize WGCNA module eigenfeatures}
\usage{
plot_eigenfeature(x)
}
\arguments{
\item{x}{output of \code{\link{run_WGCNA}}.}
}
\description{
Create scatterplots of each module eigenfeature (principal
eigenvectors).
}
