% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ORA.R
\name{plot_ORA}
\alias{plot_ORA}
\title{Plot module ORA results}
\usage{
plot_ORA(
  x,
  n_terms = 5,
  mods = 1:5,
  gs_subcat = "GO:MF",
  rel_heights = c(0.8, 0.2)
)
}
\arguments{
\item{x}{output of \code{\link[fgsea]{fora}} or \code{\link{fora2}}.}

\item{n_terms}{integer; maximum number of top over-represented terms to
display from each module.}

\item{mods}{integer; the module numbers to include in the plot. By default,
terms from modules 1-5 (the 5 largest modules) are shown.}

\item{gs_subcat}{character; the gene set subcategory of terms to plot.
Default is "GO:MF".}

\item{rel_heights}{numeric; vector of length 2 specifying the relative
heights of the plot and color legend, respectively. Default is c(0.8, 0.2).}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Dotplot of the top over-represented terms by WGCNA module.
}
