% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pre_post.R
\name{plot_pre_post}
\alias{plot_pre_post}
\title{Plot pre and post-training values}
\usage{
plot_pre_post(x, pre, post, stats = NULL, width = 0.8, angle = 20, scale = 1)
}
\arguments{
\item{x}{\code{data.frame} with pre and post values. The data should be
filtered to a specific age and sex.}

\item{pre}{character; name of a column in \code{x} containing pre-training
values.}

\item{post}{character; name of a column in \code{x} containing post-training
values.}

\item{stats}{\code{data.frame} with post - pre stats for each training group.
The data should be filtered to a specific measure, age, and sex.}

\item{width}{numeric; value between 0 and 1 specifying the maximum width of
the group with the most samples.}

\item{angle}{numeric; angle of the arrow heads.}

\item{scale}{numeric; scaling factor for text and other plot elements.}
}
\value{
A \code{ggplot} object.
}
\description{
Create a \code{ggplot2} object with an arrow for each sample
that spans from its pre-training to post-training value.
}
