% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_upset.R
\name{plot_upset}
\alias{plot_upset}
\title{Create Annotated UpSet Plot}
\usage{
plot_upset(
  x,
  top_n_comb = 10,
  comb_title = "Intersection\\nSize",
  set_title = "Set Size",
  scale = 2,
  barplot_args = list(),
  filename = character(0),
  height = 2 * scale,
  width = 3.4 * scale,
  units = "in",
  save_args = list(),
  upset_args = list()
)
}
\arguments{
\item{x}{list; named list of elements in each set. Passed to
\code{\link[ComplexHeatmap]{make_comb_mat}}.}

\item{top_n_comb}{numeric; number of largest intersections to show. Default
is 10.}

\item{comb_title}{character; title of top barplot annotation. Default is
"Intersection Size".}

\item{set_title}{character; title of right barplot annotation. Default is
"Set Size".}

\item{scale}{numeric; scaling factor for all plot elements. Improves
appearance. Default is 2.}

\item{barplot_args}{list; additional arguments passed to
\code{\link[ComplexHeatmap]{anno_barplot}}.}

\item{filename}{character; the filename used to save the UpSet plot. If
provided, the plot will not be displayed.}

\item{height}{numeric; the save height.}

\item{width}{numeric; the save width.}

\item{units}{character; the units of \code{height} and \code{width}. Default is "in"
(inches).}

\item{save_args}{list; additional arguments passed to the graphics device.
See \code{\link[grDevices]{png}} for options.}

\item{upset_args}{list; additional arguments passed to
\code{\link[ComplexHeatmap]{UpSet}}.}
}
\description{
Create an UpSet plot annotated with intersection and set sizes.
Essentially a modified version of \code{\link[ComplexHeatmap]{UpSet}}.
}
\details{
See \code{\link[ComplexHeatmap]{UpSet}} for details and sections
mentioned in Arguments.
}
\references{
Lex, A., Gehlenborg, N., Strobelt, H., Vuillemot, R., & Pfister,
H. (2014). UpSet: Visualization of Intersecting Sets. \emph{IEEE transactions on
visualization and computer graphics, 20}(12), 1983--1992.
\url{https://doi.org/10.1109/TVCG.2014.2346248}
}
