% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Create Volcano Plot}
\usage{
plot_volcano(x, pval_cutoff = 0.05, colors = c("#3366ff", "darkred", "grey"))
}
\arguments{
\item{x}{a \code{matrix} or \code{data.frame} containing differential
analysis results. Must include columns logFC, log10_pval, sign_logFC,
contrast, and label.}

\item{pval_cutoff}{numeric; cutoff for p-values to be considered significant.
Adds a dashed horizontal line to the plot.}

\item{colors}{character; length 3 vector of significantly negative,
significantly positive, and non-significant (NS) logFC colors.}
}
\value{
A \code{ggplot2} object.
}
\description{
Create a volcano plot.
}
