% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_hist.R
\name{pval_hist}
\alias{pval_hist}
\title{Create p-value Histograms}
\usage{
pval_hist(x, pval = "adj.P.Val", ylims = c(0, NA))
}
\arguments{
\item{x}{a \code{matrix} or \code{data.frame} containing differential
analysis results. Must include columns logFC, log10_pval, sign_logFC,
contrast, and label.}

\item{pval}{character; the name of a column in \code{x} containing p-values
or adjusted p-values.}

\item{ylims}{numeric vector of length 2. Limits of y-axis.}
}
\value{
A \code{ggplot2} object.
}
\description{
Create p-value histograms with
  \code{\link[ggplot2]{ggplot2-package}}.
}
