% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_extend.R
\name{range_extend}
\alias{range_extend}
\title{Extended Range of Values}
\usage{
range_extend(x, nearest = 1)
}
\arguments{
\item{x}{any \code{\link[base]{numeric}} object.}

\item{nearest}{numeric; the range of \code{x} will be rounded out to the
value specified by \code{nearest}. Default is 1.}
}
\value{
A \code{\link[base]{numeric}} vector of length 2 specifying the
  range of values after rounding outward to the value provided by
  \code{nearest}.
}
\description{
\code{range_extend} returns a vector containing the minimum and
  maximum of all the given arguments rounded outward to the value provided by
  \code{nearest}.
}
\examples{
set.seed(0)
x <- runif(5, min = -10, max = 10)
range(x) # -4.689827  8.164156

range_extend(x) # -5  9
range_extend(x, nearest = 0.1) # -4.7  8.2
}
\seealso{
\code{\link[base]{range}}
}
\author{
Tyler Sagendorf
}
