% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_genes.R
\name{rank_genes}
\alias{rank_genes}
\title{Calculate GSEA ranking metric values}
\usage{
rank_genes(
  x,
  genes,
  metric = "-log10(P.Value)*sign(logFC)",
  contrast_column = "contrast"
)
}
\arguments{
\item{x}{object of class \code{data.frame}. Typically results from
\code{\link{limma_full}}.}

\item{genes}{character; the name of a column in \code{x}. A ranking metric
value will be calculated for each unique entry. Usually gene symbols,
Entrez IDs, or Ensembl genes to work with the output of \code{msigdbr2}.}

\item{metric}{character; specifies how the ranking metric will be calculated
from \code{x}. Can be the name of a single column.}

\item{contrast_column}{character; the name of a column in \code{x}. Entries
in this column will be set as the names of the output list.}
}
\value{
A named vector (if no "contrast" column) or list of named vectors of
sorted ranking metric values for each contrast in \code{x}.
}
\description{
Calculate GSEA ranking metric values for \code{\link{fgsea2}}.
}
\details{
The -log\eqn{_{10}}-transformed p-value signed by the fold-change is
similar to the t-statistic, but more clearly separates extreme values in
the tails from those close to 0 that are not as interesting.
}
