% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pub.R
\name{theme_pub}
\alias{theme_pub}
\title{Custom ggplot2 theme}
\usage{
theme_pub()
}
\description{
Theme used extensively by the plotting functions of this
  package, as well as to generate final plots for the associated publication.
  Built to adhere to the Nature publishing guidelines.
}
