% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_GO_names.R
\name{update_GO_names}
\alias{update_GO_names}
\title{Update Gene Ontology term descriptions}
\usage{
update_GO_names(
  x,
  version = packageVersion("msigdbr"),
  capitalize = FALSE,
  obo_file
)
}
\arguments{
\item{x}{object of class \code{data.frame} produced by
\code{link[msigdbr]{msigdbr}} or \code{\link{msigdbr2}} containing columns
\code{gs_description} and \code{gs_subcat}.}

\item{version}{character; specifies the version of \code{msigdbr} to use.
Defaults to the current version.}

\item{capitalize}{logical; whether to capitalize the first letter of each
description if the first word does not contain a mix of capital and
lowercase letters. Improves appearance of plots, such as those produced by
\code{\link{enrichmat}}.}

\item{obo_file}{character; optional path to the OBO file that should be used
to update Gene Ontology term descriptions. Only provide when the
appropriate OBO file cannot be extracted from the MSigDB Release Notes (See
Details).}
}
\value{
Object of class \code{data.frame}. The same as \code{x}, but with
updated descriptions.
}
\description{
Update entries in the \code{gs_description} column of
\code{\link{msigdbr}} results to use terms from the
\href{http://geneontology.org/}{Gene Ontology Consortium}.
}
\details{
This function assumes that the phrase "GO-basic obo file released
on" is present in the MSigDB release notes for that version and is followed
by a date. This date will replace "RELEASE_DATE" in
"http://release.geneontology.org/RELEASE_DATE/ontology/go-basic.obo".
}
\examples{
x <- msigdbr2(species = "Homo sapiens",
              genes = "gene_symbol",
              gs_subcat = "GO:MF")
set.seed(9900)
idx <- sample(1:nrow(x), size = 6) # random indices to illustrate changes
x$gs_description[idx] # before

y <- update_GO_names(x, capitalize = TRUE)
y$gs_description[idx] # after
}
\references{
Ashburner, M., et al. (2000). Gene ontology: tool for the
unification of biology. The Gene Ontology Consortium. \emph{Nature
genetics, 25}(1), 25--29. \url{https://doi.org/10.1038/75556}

Gene Ontology Consortium (2021). The Gene Ontology resource: enriching a
GOld mine. \emph{Nucleic acids research, 49}(D1), D325--D334.
\url{https://doi.org/10.1093/nar/gkaa1113}
}
\seealso{
\href{https://software.broadinstitute.org/cancer/software/gsea/wiki/index.php/Release_Notes}{MSigDB
Release Notes}, \href{http://release.geneontology.org/}{Gene Ontology Data
Archive}
}
\author{
Tyler Sagendorf
}
