% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fora2.R
\name{fora2}
\alias{fora2}
\title{Over-representation analysis}
\usage{
fora2(
  pathways,
  genes,
  universe,
  gene_column = "entrez_gene",
  minSize = 1,
  maxSize = length(universe) - 1,
  adjust.method = "scale",
  adjust.globally = TRUE
)
}
\arguments{
\item{pathways}{output of \code{link{msigdbr2}}.}

\item{genes}{named list of genes. These are usually clusters of related
genes.}

\item{universe}{character vector of background genes. Usually, this is all
unique elements of \code{genes}.}

\item{gene_column}{character; the name of the column in pathways containing
the elements of each gene set.}

\item{minSize}{integer; minimum size of gene sets allowed for testing.}

\item{maxSize}{integer; maximum size of gene sets allowed for testing.}

\item{adjust.method}{character string; p-value correction method. Either
"scale" (default), which will compute scaled p-values or one of
\code{\link[stats]{p.adjust.methods}}.}

\item{adjust.globally}{logical; should p-values from all contrasts be
adjusted together using \code{adjust.method}? Set to \code{FALSE} if the
contrasts being tested are not closely related.}
}
\description{
Custom wrapper for \code{\link[fgsea]{fora}}.
}
\details{
If \code{adjust.method = "scale"}, the function will calculate the
maximum overlap for each combination of \code{names(genes)} and (if
\code{adjust.globally = TRUE}), the entries of the "gs_subcat" column of
\code{pathways}. For each gene set, it will then calculate the ratio of the
size of that set's overlap to one plus the maximum overlap. The addition of
1 in the denominator is to penalize small maximum overlaps. The
\code{log10(pval)} is multiplied by this overlap ratio and then
back-transformed to obtain p-values adjusted by how well they describe each
group defined by \code{genes}.
}
\seealso{
\code{\link[fgsea]{fora}}
}
\author{
Tyler Sagendorf
}
