% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotMDS.R
\name{ggplotMDS}
\alias{ggplotMDS}
\title{MDS plots with ggplot2}
\usage{
ggplotMDS(x)
}
\arguments{
\item{x}{an \code{link[MSnbase]{MSnSet}} object.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
A `ggplot2` version of \code{\link[limma]{plotMDS}} tailored to
  the MoTrPAC PASS1B data. Samples are labeled by the number of weeks of
  exercise training (0, 1, 2, 4, 8) and colored according to sex.
}
