% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_WGCNA.R
\name{run_WGCNA}
\alias{run_WGCNA}
\title{Run WGCNA Pipeline}
\usage{
run_WGCNA(
  object,
  power = 1:20,
  RsquaredCut = 0.9,
  module_prefix = "",
  merge_modules
)
}
\arguments{
\item{object}{object of class \code{\link[MSnbase:MSnSet-class]{MSnSet}}.
\code{exprs} should be a matrix of log\eqn{_2}-transformed values. Count
data is not accepted.}

\item{power}{integer; optional soft power vector (length 1 or greater). If
not specified, the lowest value of 1:20 that satisfies scale-free fit
R\eqn{^2} at least equal to \code{RsquaredCut} will be used.}

\item{RsquaredCut}{numeric; desired minimum scale-free fit R\eqn{^2}. Ignored
if \code{power} is provided.}

\item{module_prefix}{character; what to append to the module numbers to
create the "moduleID" column in \code{fData(object)}.}

\item{merge_modules}{logical; if not provided, user input is requested to
determine whether highly-correlated (r=0.85) modules should be combined
based on plots that the function creates. Must be provided if session is
not interactive.}
}
\value{
Object of class \code{\link[MSnbase:MSnSet-class]{MSnSet}} with
  modified \code{pData} and \code{fData}. The \code{pData} will include
  columns for each of the module eigenfeatures (prefixed with "ME" and
  followed by a color), while the \code{fData} will include columns
  "moduleColor" and "moduleID" that specify the modules to which each feature
  belongs. Features from the "grey" module should be ignored.
}
\description{
Run the WGCNA pipeline used in the WAT manuscript.
}
\references{
Zhang, B., & Horvath, S. (2005). A general framework for weighted
  gene co-expression network analysis. \emph{Statistical applications in
  genetics and molecular biology, 4}, Article17.
  \url{https://doi.org/10.2202/1544-6115.1128}

  Langfelder P and Horvath S, WGCNA: an R package for weighted correlation
  network analysis. BMC Bioinformatics 2008, 9:559
  \url{https://doi.org/10.1186/1471-2105-9-559}

  Peter Langfelder, Steve Horvath (2012). Fast R Functions for Robust
  Correlations and Hierarchical Clustering. \emph{Journal of Statistical
  Software, 46}(11), 1--17. \url{http://www.jstatsoft.org/v46/i11/}

  Horvath, S. (2011). Weighted Network Analysis. \emph{Springer New York}.
  \url{https://doi.org/10.1007/978-1-4419-8819-5}

  Langfelder P, Zhang B, Horvath wcfS (2016). \emph{dynamicTreeCut: Methods
  for Detection of Clusters in Hierarchical Clustering Dendrograms}. R
  package version 1.63-1,
  \url{https://CRAN.R-project.org/package=dynamicTreeCut}.
}
\author{
Tyler Sagendorf, Zhenxin Hou
}
