% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_GO_names.R
\name{update_GO_names}
\alias{update_GO_names}
\title{Update Gene Ontology term descriptions}
\usage{
update_GO_names(x, version = packageVersion("msigdbr"), capitalize = FALSE)
}
\arguments{
\item{x}{a `data.frame` produced by \code{link[msigdbr]{msigdbr}} or
\code{\link{msigdbr2}} containing columns `gs_description` and `gs_subcat`.}

\item{version}{character string specifying the version of `msigdbr` to use.
Defaults to the current version.}

\item{capitalize}{logical; whether to capitalize the first letter of each
description if the first word does not contain a mix of capital and
lowercase letters. Improves appearance of plots, such as those produced by
\code{\link{enrichmat}}.}
}
\value{
A `data.frame`. The same as `x`, but with updated descriptions.
}
\description{
Update entries in the `gs_description` column of
  \code{\link{msigdbr}} results to use terms from the
  \href{http://geneontology.org/}{Gene Ontology Consortium}.
}
\details{
This function assumes that the phrase "GO-basic obo file released
  on" is present in the MSigDB release notes for that version and is followed
  by a date.
}
\examples{
x <- msigdbr2(species = "Homo sapiens",
              genes = "gene_symbol",
              gs_subcat = "GO:MF")
tail(x$gs_description) # before

x <- update_GO_names(x)
tail(x$gs_description) # after
}
\references{
Ashburner, M., et al. (2000). Gene ontology: tool for the
  unification of biology. The Gene Ontology Consortium. *Nature genetics*,
  *25*(1), 25–29. \url{https://doi.org/10.1038/75556}

  Gene Ontology Consortium (2021). The Gene Ontology resource: enriching a
  GOld mine. *Nucleic acids research*, *49*(D1), D325–D334.
  \url{https://doi.org/10.1093/nar/gkaa1113}
}
\seealso{
\href{https://software.broadinstitute.org/cancer/software/gsea/wiki/index.php/Release_Notes}{MSigDB
Release Notes} \href{http://release.geneontology.org/}{Gene Ontology Data
Archive}
}
\author{
Tyler Sagendorf
}
